move resolution transforms inside the editor
The canvases are currently only scaled up during renders which works fine
for the flood filling logic, but not for the logic where we sample the color
initially to decide to skip the flood fill operation.

This change just ensures that the utils are generic and that the editor
is always responsible for applying the resolution to coordinates whenever
it needs to work directly with pixels.